Move calculateTransformOrigin to helpers and use it in person poster too
