+++ src/lib/Poster.svelte
@@ -1,7 +1,12 @@
 <script lang="ts">
   import type { MediaType, WatchedStatus } from "@/types";
   import Icon from "./Icon.svelte";
-  import { addClassToParent, isTouch, watchedStatuses } from "@/lib/util/helpers";
+  import {
+    addClassToParent,
+    calculateTransformOrigin,
+    isTouch,
+    watchedStatuses
+  } from "@/lib/util/helpers";
   import { goto } from "$app/navigation";
   import tooltip from "./actions/tooltip";
 
@@ -42,38 +47,6 @@
       return;
     }
     onStatusChanged(type);
-  }
-
-  function calculateTransformOrigin(
-    e: Event & {
-      currentTarget: EventTarget & HTMLLIElement;
-    }
-  ) {
-    const magicNumber = 26;
-    const ctr = e.currentTarget.querySelector(".container") as HTMLElement;
-    const pb = ctr.getBoundingClientRect();
-    const sx = pb.x;
-    const sw = pb.width;
-    const wb = document.body.getBoundingClientRect();
-
-    if (ctr) {
-      ctr.style.transformOrigin = "unset";
-      let origins = [];
-      // Overflow on right
-      if (sx + sw + magicNumber > wb.x + wb.width) {
-        origins.push("right");
-      }
-      // Overflow on left
-      if (sx - magicNumber < wb.x) {
-        origins.push("left");
-      }
-      // Overflow on bottom
-      const ppb = e.currentTarget.getBoundingClientRect();
-      if (ppb.bottom + magicNumber > window.innerHeight) {
-        origins.push("bottom");
-      }
-      ctr.style.transformOrigin = `${origins.join(" ")}`;
-    }
   }
 </script>
